Moved isGenericMemberAccessName into utils package.
