+++ query-primitives/src/main/scala/io/shiftleft/queryprimitives/utils/MemberAccess.scala
@@ -1,0 +1,20 @@
+package io.shiftleft.queryprimitives.utils
+
+import io.shiftleft.codepropertygraph.generated.Operators
+
+object MemberAccess {
+
+  /**
+    * For a given name, determine whether it is the
+    * name of a "member access" operation, e.g.,
+    * "<operator>.memberAccess".
+    * */
+  def isGenericMemberAccessName(name: String): Boolean = {
+    (name == Operators.memberAccess) ||
+    (name == Operators.indirectComputedMemberAccess) ||
+    (name == Operators.indirectMemberAccess) ||
+    (name == Operators.computedMemberAccess) ||
+    (name == Operators.indirection)
+  }
+
+}
